Move `get_css` to `Style`
Up till now `get_css` was a separate request. This takes time and is not
efficient. Since `Style` has the css, it is now used to render it
ourselves. It uses the default settings and allows user overrides, so it
should work the same exact way. Because the requests are the main
overheads this has almost doubled speed of `get_css`.

All changes have been documented and tests have been created or
rewritten to deal with it.

Also incremented version from `0.1.0` to `0.1.1`.