Fik major bug in get_css
Until now `get_css` assumed the templates in the CSS all start with
`ik-`, similar to the way the requests to the API work.
However this is not the case and lead to CSS variable replacement
being broken all of the time.

A test has been added to make sure this does not happen again.